+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Omega Strike` Special power.
 - Added `Summon Beast` Special power.
 - Added `Armor Saint` Special power.
+- Added `True Resurrection` Overpowered ability.
 - Added `Playful Darkness` to `Summon Calamity` Overpowered ability.
 - Added bard, ranger, paladin, and inquisitor spells into the `Isekai Protagonist` spellbook.
 - Added `Mythic Special Power` to mythic ability selection.